State: Move withEnhancers() to a more appropriate place
